Add ReEncryptDeviceQueueItems method on AS interface.
As the LoRaWAN 1.0.4 specs states that mac-commands have priority over
application payloads and because mac-layer only downlinks share the same
frame-counter for 1.0.x devices, the NS needs an AS api for
re-encryption of application payloads. Note that the FCnt is one of the
parameters of the FRMPayload encryption scheme.

Example:

Application payload encrypted with FCnt 10 exists in the queue. The NS
must send a block of mac-commands to the device and uses FCnt 10 for the
downlink, because of max. payload restrictions it is not able to send
the application payload. As the next downlink will have FCnt 11, it will
request the AS to re-encrypt the FRMPayload with FCnt 10 using FCnt 11,
so that it can be sent at the next downlink opportunity.